feat: moved helpers out
